# training/26ChiRouter/docs/swagger.yaml
basePath: /
definitions:
  models.Book:
    properties:
      author:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      price:
        description: gt is range validate
        type: integer
      title:
        maxLength: 50
        minLength: 2
        type: string
      updatedAt:
        type: string
      year:
        type: integer
    required:
    - author
    - id
    - price
    - title
    - year
    type: object
host: localhost:3000
info:
  contact: {}
  description: This is a sample Book API using Chi Router.
  title: Book API
  version: "1.0"
paths:
  /books:
    get:
      consumes:
      - application/json
      description: get all books in the store
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Book'
            type: array
      summary: Get all books
      tags:
      - books
    post:
      consumes:
      - application/json
      description: add a new book to the store
      parameters:
      - description: Book object
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.Book'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Invalid JSON payload
          schema:
            type: string
      summary: Create a new book
      tags:
      - books
  /books/{id}:
    delete:
      consumes:
      - application/json
      description: delete book by ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID format
          schema:
            type: string
      summary: Delete a book
      tags:
      - books
    get:
      consumes:
      - application/json
      description: get book by ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Invalid ID format
          schema:
            type: string
        "404":
          description: Book not found
          schema:
            type: string
      summary: Get a book by ID
      tags:
      - books
    put:
      consumes:
      - application/json
      description: update book by ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Book object
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.Book'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Invalid JSON payload
          schema:
            type: string
        "404":
          description: Book not found
          schema:
            type: string
      summary: Update an existing book
      tags:
      - books
swagger: "2.0"